+++ src/core/plugins/core-barbershop/webview/Barbershop.vue
@@ -101,7 +101,7 @@ export default defineComponent({
             // Update Check
             ready: false,
             // Navigation
-            navIndex: 2,
+            navIndex: 0,
             navigation: [
                 { icon: 'icon-hair', type: 'hair', colors: 2, isHair: true, colorComponentType: 0 },
                 { icon: 'icon-eye', type: 'eye', colors: 1, isEyebrows: true, colorComponentType: 0 },
@@ -192,8 +192,6 @@ export default defineComponent({
 
             for (const key of Object.keys(data)) {
                 this[key] = data[key];
-
-                console.log(`${key}: ${data[key]}`);
             }
 
             this.update();
@@ -205,6 +203,10 @@ export default defineComponent({
                 if (navIndex !== -1) {
                     navigationOptions.splice(navIndex, 1);
                     this.navigation = navigationOptions;
+                }
+
+                if (navIndex >= this.navigation.length) {
+                    this.navIndex = 0;
                 }
             }