+++ src/core/plugins/core-hud/webview/components/Fuel.vue
@@ -2,11 +2,11 @@
     <div class="stat-wrapper">
         <div class="split space-between">
             <div class="bar">
-                <div class="bar-shadow"></div>
-                <Icon :shadow="true" class="stat-icon" icon="icon-shield2" :size="18"></Icon>
+                <Icon :shadow="false" class="stat-icon" icon="icon-local_gas_station" :size="18"></Icon>
                 <div class="fill" :style="getFill">
-                    <div class="fill-shine" :style="getFill" />
+                    <img class="oil-slick" :src="ResolvePath(`../../assets/images/oil.png`)" />
                 </div>
+                <div class="bar-shadow"></div>
             </div>
         </div>
     </div>
@@ -14,8 +14,9 @@
 
 <script lang="ts">
 import { defineComponent, defineAsyncComponent } from 'vue';
+import ResolvePath from '@utility/pathResolver';
 
-const ComponentName = 'Armour';
+const ComponentName = 'Fuel';
 export default defineComponent({
     name: ComponentName,
     components: {
@@ -24,11 +25,18 @@ export default defineComponent({
     props: {
         value: {
             type: Number,
-            default: 199,
+            default: 50,
         },
+    },
+    methods: {
+        ResolvePath,
     },
     computed: {
         getFill() {
+            if (this.value >= 99) {
+                return `height: ${this.value}% !important; animation: unset !important;`;
+            }
+
             return `height: ${this.value}% !important`;
         },
     },
@@ -86,46 +94,45 @@ export default defineComponent({
     z-index: 99;
     opacity: 0.5;
     left: -4px;
-    top: -1px;
+    top: -2px;
     mix-blend-mode: lighten;
 }
 
 .fill {
+    display: flex;
     position: absolute;
-    background: #475e5e;
+    background: transparent;
     z-index: 25;
     min-width: 65px;
+    max-width: 65px;
     box-sizing: border-box;
     bottom: 0px;
+    animation: rotate 5s infinite ease-in-out;
     outline: 1px solid transparent;
-    overflow: hidden;
+    overflow: hidden !important;
+    justify-content: center;
+    align-items: center;
 }
 
-.fill-shine {
-    position: relative;
-    background: #87b3b3;
-    z-index: 30;
-    min-width: 2px;
-    max-width: 2px;
-    min-height: 100px;
-    box-sizing: border-box;
-    top: -25px;
-    animation: swipe 4s infinite linear;
-    box-shadow: 0px 0px 10px 10px #87b3b3;
+.oil-slick {
+    width: 150%;
+    height: 150%;
+    animation: rotate 7s infinite alternate;
+    object-fit: cover;
 }
 
-@keyframes swipe {
+@keyframes rotate {
     0% {
-        transform: translateX(-100px) rotate(45deg);
-        opacity: 0;
+        transform: translateY(0px);
     }
     25% {
-        transform: translateX(40px) rotate(45deg);
-        opacity: 0.8;
+        transform: translateY(2px) rotate(-5deg);
+    }
+    75% {
+        transform: translateY(2px) rotate(5deg);
     }
     100% {
-        transform: translateX(200px) rotate(45deg);
-        opacity: 1;
+        transform: translateY(0px);
     }
 }
 </style>