+++ src/core/server/extensions/playerFunctions/shared.ts
@@ -1,0 +1,4 @@
+import * as alt from 'alt-server';
+import { Character } from '../../../shared/interfaces/character';
+
+export type ReadOnlyPlayer = Readonly<alt.Player & { data: Character }>;