+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## 3.7.0
+
+```
+Fix Barbershop Warnings
+Add Catch for Camera Destroy Failure for Barbershop
+Remove Barbershop Console Logs
+```
+
+## 3.6.0
+
+```
+Refactoring for Strict Typing Support (WIP)
+Changed Icon !important Strictness in src-webviews
+Fix Female Barbershop Error
+Redesigned Core HUD
++ Added Player ID
++ Added F2 to Cycle HUD Displays
++ Added Death State to Core HUD Updates
++ Added Player ID to Core HUD Updates
+```
+
 ## 3.5.0
 
 ```